+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.0 - 2024-12-02
+### Added
+- Add support for custom setup scripts
+- Add terminal notifier to get notified of update result
+
 ## 1.0.1 - 2024-10-21
 ### Changed
 - Enable automatic deployments